+++ lib/src/test_models.dart
@@ -61,22 +61,22 @@ class QueryByName extends Query {
   final String namePrefix;
 }
 
-class QueryByCreatedAfter extends Query {
-  const QueryByCreatedAfter(this.date);
-  final DateTime date;
-}
+class QueryByPriceRange extends Query {
+  const QueryByPriceRange({this.minPrice, this.maxPrice});
+  final double? minPrice;
+  final double? maxPrice;
 
-class QueryByCreatedBefore extends Query {
-  const QueryByCreatedBefore(this.date);
-  final DateTime date;
-}
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is QueryByPriceRange &&
+          runtimeType == other.runtimeType &&
+          minPrice == other.minPrice &&
+          maxPrice == other.maxPrice;
 
-class QueryByPriceGreaterThan extends Query {
-  const QueryByPriceGreaterThan(this.price);
-  final double price;
-}
+  @override
+  int get hashCode => minPrice.hashCode ^ maxPrice.hashCode;
 
-class QueryByPriceLessThan extends Query {
-  const QueryByPriceLessThan(this.price);
-  final double price;
+  @override
+  String toString() => 'QueryByPriceRange(minPrice: $minPrice, maxPrice: $maxPrice)';
 }